double precision, new struct

+++ gaus.h
@@ -9,8 +9,8 @@
  * overwrites contents of eta, phi, pt vectors with jet constituents
  */
 void gausJet(unsigned nPart, jet& jetout,
-                             float ptscale=1,
-                             float angscale=0.4,
-                             float ptsmear=0.1);
+                             double ptscale=1,
+                             double angscale=0.4,
+                             double ptsmear=0.1);
 
 #endif
